+++ lib/heic_to_png_jpg.dart
@@ -2,3 +2,6 @@ library heic_to_png_jpg;
 
 export 'src/heic_to_png_jpg_base.dart';
 export 'src/image_format.dart';
+export 'src/exceptions.dart';
+export 'src/heic_image_info.dart';
+export 'src/heic_conversion_task.dart';