+++ lib/src/image_format.dart
@@ -1,4 +1,5 @@
 enum ImageFormat {
   jpg,
   png,
+  webp,
 }